Re-organized range mapping functionality
