+++ woltka/range.py
@@ -22,6 +22,76 @@ Filter program (https://github.com/ucsd-cmi/zebra_filter).
 from os import makedirs
 from os.path import join
 
+from .align import iter_align
+
+
+def range_mapper(fh, fmt=None, excl=None, n=1000):
+    """Read an alignment file and yield maps of query to subject(s) and their
+    ranges.
+
+    Parameters
+    ----------
+    fh : file handle
+        Alignment file to parse.
+    fmt : str, optional
+        Alignment file format.
+    excl : set, optional
+        Subjects to exclude.
+    n : int, optional
+        Number of unique queries per chunk.
+
+    Yields
+    ------
+    list of str
+        Query queue.
+    list of dict of list of int
+        Subject-to-ranges queue.
+
+    Notes
+    -----
+    Same as `plain_mapper`, except that it also returns subject ranges.
+
+    Ranges are stored as a one-dimensional, interleaved list of start1, end1,
+    start2, end2, start3, end3...
+
+    See Also
+    --------
+    .align.plain_mapper
+    merge_ranges
+    """
+    it = iter_align(fh, fmt, excl, True)
+    while True:
+        i, done = 0, False
+        qryque, subque = [None] * n, [None] * n
+        for query, records in it:
+
+            # generate a mapping of subjects to interleaved starts and ends
+            ranges = {}
+            for subject, _, _, start, end in records:
+
+                # start and end must be positive integers
+                if start and end:
+
+                    # combine ranges on the same subject
+                    ranges.setdefault(subject, []).extend((start, end))
+
+            # append query and ranges
+            if ranges:
+                qryque[i] = query
+                subque[i] = ranges
+
+                i += 1
+                if i == n:
+                    done = True
+                    break
+
+        if not done:
+            if i:
+                yield qryque[:i], subque[:i]
+            break  # pragma: no cover
+        else:
+            yield qryque, subque
+
 
 def merge_ranges(ranges):
     """Merge short, fragmental ranges into long, continuous ones.
